+++ general/thread_system.h
@@ -52,7 +52,7 @@ namespace System
 		#endif
 	}
 
-	#define THREAD_REPORT(message)	if (System::report_kernel_activity){System::thread_report_mutex.lock(); std::cout << ((void*)System::thisThread())<<" (this="<<this<<")<"<<__func__<<">: "<<message<<std::endl;  System::thread_report_mutex.release();}
+	#define THREAD_REPORT(message)	if (System::report_kernel_activity){System::thread_report_mutex.lock(); std::cout << ((void*)(size_t)System::thisThread())<<" (this="<<this<<")<"<<__func__<<">: "<<message<<std::endl;  System::thread_report_mutex.release();}
 	
 
 	class ThreadMainObject;
